Move CacheDispatcher tests into their own file.
