+++ tests/TestCase/Routing/RouterTest.php
@@ -1,7 +1,5 @@
 <?php
 /**
- * RouterTest file
- *
  * CakePHP(tm) Tests <http://book.cakephp.org/2.0/en/development/testing.html>
  * Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
  *
